Move IntegrationTest to a separate module to avoid akka-grpc version conflict
