+++ integration-test/src/test/proto/cloudstate/crdt.proto
