+++ integration-test/src/test/proto/cloudstate/entity.proto
