+++ integration-test/src/test/proto/cloudstate/entity_key.proto
