+++ integration-test/src/test/proto/cloudstate/event_sourced.proto
