+++ integration-test/src/test/proto/cloudstate/eventing.proto
