+++ integration-test/src/test/proto/cloudstate/function.proto
