+++ integration-test/src/test/proto/google/api/annotations.proto
