+++ integration-test/src/test/proto/google/api/http.proto
