+++ integration-test/src/test/proto/google/api/httpbody.proto
