+++ integration-test/src/test/proto/google/protobuf/any.proto
