+++ integration-test/src/test/proto/google/protobuf/descriptor.proto
