+++ integration-test/src/test/proto/google/protobuf/empty.proto
