+++ integration-test/src/test/proto/shoppingcart.proto
