+++ settings.gradle.kts
@@ -4,3 +4,4 @@ include(":cloudstate-kotlin-support")
 include(":examples:kotlin-chat")
 include(":examples:kotlin-pingpong")
 include(":examples:shopping-cart")
+include(":integration-test")
